+++ src/main/java/org/moon/figura/avatars/model/rendering/texture/FiguraTexture.java
@@ -12,7 +12,7 @@ import org.lwjgl.BufferUtils;
 import org.moon.figura.FiguraMod;
 import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.docs.LuaMethodDoc;
-import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
+import org.moon.figura.lua.docs.LuaMethodOverload;
 import org.moon.figura.lua.docs.LuaTypeDoc;
 import org.moon.figura.math.vector.FiguraVec2;
 import org.moon.figura.math.vector.FiguraVec3;
