+++ src/main/java/org/moon/figura/lua/api/TextureAPI.java
@@ -8,7 +8,7 @@ import org.moon.figura.avatars.model.rendering.texture.FiguraTextureSet;
 import org.moon.figura.lua.LuaNotNil;
 import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.docs.LuaMethodDoc;
-import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
+import org.moon.figura.lua.docs.LuaMethodOverload;
 import org.moon.figura.lua.docs.LuaTypeDoc;
 import org.moon.figura.utils.ColorUtils;
 
@@ -69,7 +69,7 @@ public class TextureAPI {
 
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaFunctionOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = {String.class, String.class},
                     argumentNames = {"name", "data"}
             ),
