+++ src/main/java/org/moon/figura/lua/api/entity/LivingEntityAPI.java
@@ -7,7 +7,7 @@ import net.minecraft.world.entity.LivingEntity;
 import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.api.world.ItemStackAPI;
 import org.moon.figura.lua.docs.LuaMethodDoc;
-import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
+import org.moon.figura.lua.docs.LuaMethodOverload;
 import org.moon.figura.lua.docs.LuaTypeDoc;
 
 import java.util.ArrayList;
