+++ src/main/java/org/moon/figura/lua/api/event/LuaEvent.java
@@ -9,6 +9,7 @@ import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.docs.LuaMetamethodDoc;
 import org.moon.figura.lua.docs.LuaMetamethodDoc.LuaMetamethodOverload;
 import org.moon.figura.lua.docs.LuaMethodDoc;
+import org.moon.figura.lua.docs.LuaMethodOverload;
 import org.moon.figura.lua.docs.LuaTypeDoc;
 
 import java.util.Deque;
@@ -63,11 +64,11 @@ public class LuaEvent {
     @LuaWhitelist
     @LuaMethodDoc(
             overloads = {
-                    @LuaMethodDoc.LuaMethodOverload(
+                    @LuaMethodOverload(
                             argumentTypes = LuaFunction.class,
                             argumentNames = "func"
                     ),
-                    @LuaMethodDoc.LuaMethodOverload(
+                    @LuaMethodOverload(
                             argumentTypes = {LuaFunction.class, String.class},
                             argumentNames = {"func", "name"}
                     )
@@ -93,7 +94,7 @@ public class LuaEvent {
 
     @LuaWhitelist
     @LuaMethodDoc(
-            overloads = @LuaMethodDoc.LuaMethodOverload(
+            overloads = @LuaMethodOverload(
                     argumentTypes = String.class,
                     argumentNames = "name"
             ),
