+++ src/main/java/org/moon/figura/lua/api/math/MatricesAPI.java
@@ -3,7 +3,7 @@ package org.moon.figura.lua.api.math;
 import org.luaj.vm2.LuaError;
 import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.docs.LuaMethodDoc;
-import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
+import org.moon.figura.lua.docs.LuaMethodOverload;
 import org.moon.figura.lua.docs.LuaTypeDoc;
 import org.moon.figura.math.matrix.FiguraMat2;
 import org.moon.figura.math.matrix.FiguraMat3;
