+++ src/main/java/org/moon/figura/lua/api/nameplate/NameplateCustomizationGroup.java
@@ -3,7 +3,7 @@ package org.moon.figura.lua.api.nameplate;
 import org.moon.figura.lua.LuaNotNil;
 import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.docs.LuaMethodDoc;
-import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
+import org.moon.figura.lua.docs.LuaMethodOverload;
 import org.moon.figura.lua.docs.LuaTypeDoc;
 
 import java.util.Arrays;
