+++ src/main/java/org/moon/figura/lua/api/vanilla_model/VanillaGroupPart.java
@@ -3,7 +3,7 @@ package org.moon.figura.lua.api.vanilla_model;
 import net.minecraft.client.model.EntityModel;
 import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.docs.LuaMethodDoc;
-import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
+import org.moon.figura.lua.docs.LuaMethodOverload;
 import org.moon.figura.lua.docs.LuaTypeDoc;
 import org.moon.figura.utils.ui.UIHelper;
