+++ src/main/java/org/moon/figura/lua/api/vanilla_model/VanillaModelPart.java
@@ -5,7 +5,7 @@ import net.minecraft.client.model.geom.ModelPart;
 import org.moon.figura.avatars.model.ParentType;
 import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.docs.LuaMethodDoc;
-import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
+import org.moon.figura.lua.docs.LuaMethodOverload;
 import org.moon.figura.lua.docs.LuaTypeDoc;
 import org.moon.figura.math.vector.FiguraVec3;
 import org.moon.figura.utils.ui.UIHelper;
