+++ src/main/java/org/moon/figura/math/vector/FiguraVec3.java
@@ -10,7 +10,7 @@ import org.moon.figura.lua.LuaNotNil;
 import org.moon.figura.lua.LuaWhitelist;
 import org.moon.figura.lua.docs.*;
 import org.moon.figura.lua.docs.LuaMetamethodDoc.LuaMetamethodOverload;
-import org.moon.figura.lua.docs.LuaMethodDoc.LuaMethodOverload;
+import org.moon.figura.lua.docs.LuaMethodOverload;
 import org.moon.figura.math.matrix.FiguraMat3;
 import org.moon.figura.utils.LuaUtils;
 import org.moon.figura.utils.MathUtils;
